+++ src/lib/components/Cytoscape.react.js
@@ -214,6 +214,36 @@ class Cytoscape extends Component {
             });
         }, EXTENT_THRESHOLD);
 
+        // Store the original maxZoom and minZoom functions
+        const originalMaxZoomFn = cy.maxZoom;
+        const originalMinZoomFn = cy.minZoom;
+
+        // Override the maxZoom function to trigger maxZoomChange custom event
+        cy.maxZoom = function (e) {
+            const currentMaxZoom = originalMaxZoomFn.call(cy, e);
+
+            // Trigger your custom event if the current max zoom level is different from the
+            // previously stored max zoom level
+            if (currentMaxZoom !== cy._previousMaxZoom) {
+                cy._previousMaxZoom = currentMaxZoom;
+                cy.trigger('minMaxZoomChange');
+            }
+            return currentMaxZoom;
+        };
+
+        // Override the minZoom function to trigger minZoomChange custom event
+        cy.minZoom = function (e) {
+            const currentMinZoom = originalMinZoomFn.call(cy, e);
+
+            // Trigger your custom event if the current min zoom level is different from the
+            // previously stored min zoom level
+            if (currentMinZoom !== cy._previousMinZoom) {
+                cy._previousMinZoom = currentMinZoom;
+                cy.trigger('minMaxZoomChange');
+            }
+            return currentMinZoom;
+        };
+
         // /////////////////////////////////////// EVENTS //////////////////////////////////////////
 
         cy.on('tap', 'node', (event) => {
@@ -319,6 +349,14 @@ class Cytoscape extends Component {
 
         cy.on('viewport resize', () => {
             setExtent(cy.extent());
+        });
+
+        // Refresh layout if current zoom is out of boundaries
+        cy.on('minMaxZoomChange', function () {
+            const zoom = cy.zoom();
+            if (zoom > cy.maxZoom() || zoom < cy.minZoom()) {
+                cy.fit();
+            }
         });
 
         this.createMenuItems = (ctxMenu) => {