+++ src/lib/cyleaflet_clientside.js
@@ -10,11 +10,15 @@ if (!window.dash_clientside) {
 // Reference: https://epsg.io/3857
 
 // Conversion factor based on EPSG:3857 bounds
-var conversion_factor = 20037508.34;
+const conversion_factor = 20037508.34;
 
-var double_lon = 360;
-var max_lon = 180;
-var max_lat = 90;
+const double_lon = 360;
+const max_lon = 180;
+const max_lat = 90;
+
+const leafletZoomMultiplier = 0.418;
+const leafletZoomReference = 16;
+const defaultLeafletMaxZoom = 18;
 
 // Convert EPSG:4326 to EPSG:3857
 // We also flip the sign of the y-value to match Cytoscape's coordinate system
@@ -38,10 +42,18 @@ function xYToLonLat(x, y) {
     return [lon, lat];
 }
 
+function computeCytoscapeMaxZoom(leafletMaxZoom) {
+    return leafletZoomMultiplier * 2 ** (leafletMaxZoom - leafletZoomReference);
+}
+
 window.dash_clientside.cyleaflet = {
-    updateLeafBounds: function (cyExtent) {
+    updateLeafBounds: function (cyExtentInput, max_zoom, cyExtentStore) {
+        var cyExtent = cyExtentInput;
         if (!cyExtent) {
-            return window.dash_clientside.no_update;
+            if (!cyExtentStore) {
+                return window.dash_clientside.no_update;
+            }
+            cyExtent = cyExtentStore;
         }
 
         var lonLatMin = xYToLonLat(cyExtent.x1, cyExtent.y1);
@@ -70,6 +82,7 @@ window.dash_clientside.cyleaflet = {
                 bounds: bounds,
                 options: {animate: true},
             },
+            cyExtent,
         ];
     },
     transformElements: function (elements) {
@@ -86,5 +99,48 @@ window.dash_clientside.cyleaflet = {
             }
             return e;
         });
+    },
+    updateLonLat: function (elements) {
+        if (elements.length > 0) {
+            return elements.map((e) => {
+                if (
+                    typeof e.position !== 'undefined' &&
+                    Object.prototype.hasOwnProperty.call(e.position, 'x') &&
+                    e.position.x !== 0 &&
+                    e.position.y !== 0
+                ) {
+                    var lonLat = xYToLonLat(e.position.x, e.position.y);
+                    return {
+                        position: e.position,
+                        data: Object.assign({}, e.data, {
+                            lon: lonLat[0],
+                            lat: lonLat[1],
+                        }),
+                    };
+                }
+                return e;
+            });
+        }
+        return window.dash_clientside.no_update;
+    },
+    updateCytoMaxZoom: function (children) {
+        var tileLayer = children;
+        if (children.length >= 1) {
+            tileLayer = children.filter((e) => {
+                return e.type === 'TileLayer';
+            });
+            if (tileLayer.length === 1) {
+                tileLayer = tileLayer[0];
+            }
+        }
+
+        var leafletMaxZoom = defaultLeafletMaxZoom;
+        if (
+            typeof tileLayer.props !== 'undefined' &&
+            Object.prototype.hasOwnProperty.call(tileLayer.props, 'maxZoom')
+        ) {
+            leafletMaxZoom = tileLayer.props.maxZoom;
+        }
+        return computeCytoscapeMaxZoom(leafletMaxZoom);
     },
 };